# Conflicts:
#	src/main/java/cn/lanqiao/studentgradesmanagesystem/service/impl/resultServiceImpl.java
#	src/main/java/cn/lanqiao/studentgradesmanagesystem/service/impl/teacherServiceImpl.java

+++ src/main/java/cn/lanqiao/studentgradesmanagesystem/pojo/Result.java
@@ -4,15 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Result {
-  private int resId;
+
+  private long resId;
   private String stuId;
   private String subName;
-  private int resNum;
+  private long resNum;
   private String resTerm;
 
 }
